Move the input file to 4c
