+++ src/meshpy/four_c/dbc_monitor.py
@@ -38,7 +38,7 @@ from meshpy.core.function import Function
 from meshpy.core.function_utility import (
     create_linear_interpolation_function,
 )
-from meshpy.core.inputfile import InputFile
+from meshpy.four_c.inputfile import InputFile
 
 
 def linear_time_transformation(
