+++ tests/test_cosserat_curve.py
@@ -33,7 +33,6 @@ import quaternion
 
 from meshpy.core.conf import mpy
 from meshpy.core.element_beam import Beam3rHerm2Line3
-from meshpy.core.inputfile import InputFile
 from meshpy.core.material import MaterialReissner
 from meshpy.core.rotation import Rotation
 from meshpy.cosserat_curve.cosserat_curve import CosseratCurve
@@ -42,6 +41,7 @@ from meshpy.cosserat_curve.warping_along_cosserat_curve import (
     get_mesh_transformation,
     warp_mesh_along_curve,
 )
+from meshpy.four_c.inputfile import InputFile
 from meshpy.mesh_creation_functions.beam_basic_geometry import create_beam_mesh_helix
