+++ tests/test_four_c.py
@@ -36,11 +36,11 @@ from meshpy.core.conf import mpy
 from meshpy.core.element_beam import Beam3rHerm2Line3
 from meshpy.core.function import Function
 from meshpy.core.geometry_set import GeometrySet
-from meshpy.core.inputfile import InputFile
 from meshpy.core.material import MaterialReissner
 from meshpy.core.rotation import Rotation
 from meshpy.four_c.beam_potential import BeamPotential
 from meshpy.four_c.dbc_monitor import linear_time_transformation
+from meshpy.four_c.inputfile import InputFile
 from meshpy.four_c.locsys_condition import LocSysCondition
 from meshpy.four_c.solid_shell_thickness_direction import (
     get_visualization_third_parameter_direction_hex8,
