+++ tests/test_mesh_creation_functions.py
@@ -38,11 +38,11 @@ import splinepy
 from meshpy.core.boundary_condition import BoundaryCondition
 from meshpy.core.conf import mpy
 from meshpy.core.element_beam import Beam3eb, Beam3rHerm2Line3
-from meshpy.core.inputfile import InputFile
 from meshpy.core.material import MaterialEulerBernoulli, MaterialReissner
 from meshpy.core.mesh import Mesh
 from meshpy.core.node import NodeCosserat
 from meshpy.core.rotation import Rotation
+from meshpy.four_c.inputfile import InputFile
 from meshpy.mesh_creation_functions.beam_basic_geometry import (
     create_beam_mesh_arc_at_node,
     create_beam_mesh_arc_segment_2d,
