+++ tests/test_meshpy.py
@@ -54,7 +54,6 @@ from meshpy.core.header_functions import (
     set_header_static,
     set_runtime_output,
 )
-from meshpy.core.inputfile import InputFile, InputSection
 from meshpy.core.material import (
     MaterialBeam,
     MaterialEulerBernoulli,
@@ -67,6 +66,7 @@ from meshpy.core.mesh import Mesh
 from meshpy.core.node import Node, NodeCosserat
 from meshpy.core.rotation import Rotation
 from meshpy.core.vtk_writer import VTKWriter
+from meshpy.four_c.inputfile import InputFile, InputSection
 from meshpy.mesh_creation_functions.beam_basic_geometry import (
     create_beam_mesh_arc_segment_via_rotation,
     create_beam_mesh_line,
